requirements.txt
Django==2.1.5
djangorestframework==3.9.0
Pillow==5.3.0
django-widget-tweaks==1.4.3
mock==2.0.0
django-stdimage==4.0.0
django-suit==0.2.26
coverage==4.5.1
django-nose==1.4.6
nose==1.3.7
django-cookie-law>=0.3.0
